Update monero-wallet-generator-print.html
Removed "AeonAddress:", "spendkey:", and "privatekey:" from the QR codes. When you scan a QR code, you know what you're scanning. It is not necessary and gets in the way if you're trying to copy the address to somebody who wants to send you coins, or if you're trying to copy the private key to spend coins.